model/requirements.txt
pandas
scikit-learn
pyyaml
joblib

